+++ packages/actionsflow-core/src/utils/template.ts
@@ -304,8 +304,16 @@ export const template = function (
   }
 
   evaluate.push(stringify(text.slice(i)));
+
   // Function is needed to opt out from possible "use strict" directive
-  return Function("with(this)return " + evaluate.join("+")).call(context);
+  const functionString = evaluate
+    .map((item) => {
+      return `result += ${item};`;
+    })
+    .join("\n");
+  const finalFunctionString = `with(this) {var result="";${functionString};return result;}`;
+
+  return Function(finalFunctionString).call(context);
 };
 
 export const getTemplateStringByParentName = (
@@ -314,14 +322,17 @@ export const getTemplateStringByParentName = (
   context: AnyObject
 ): string => {
   const asts = getAstsByParentName(text, parentName);
-  let finalResults = "";
-  asts.forEach((ast) => {
-    const string = text.slice(ast.start, ast.end);
-    if (ast.type === "expression") {
-      finalResults = finalResults + getExpressionResult(string, context);
-    } else {
-      finalResults = finalResults + string;
-    }
-  });
-  return finalResults;
+  const functionString = asts
+    .map((ast) => {
+      const string = text.slice(ast.start, ast.end);
+      if (ast.type === "expression") {
+        return `result += (${string});`;
+      } else {
+        return `result += ${JSON.stringify(string)};`;
+      }
+    })
+    .join("\n");
+  const finalFunctionString = `with(this) {var result="";${functionString};return result;}`;
+
+  return Function(finalFunctionString).call(context);
 };