+++ packages/actionsflow/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.4](https://github.com/actionsflow/actionsflow/compare/actionsflow@1.5.3...actionsflow@1.5.4) (2020-11-02)
+
+**Note:** Version bump only for package actionsflow
+
+
+
+
+
+## [1.5.3](https://github.com/actionsflow/actionsflow/compare/actionsflow@1.5.2...actionsflow@1.5.3) (2020-11-02)
+
+**Note:** Version bump only for package actionsflow
+
+
+
+
+
 ## [1.5.2](https://github.com/actionsflow/actionsflow/compare/actionsflow@1.5.1...actionsflow@1.5.2) (2020-10-31)
 
 **Note:** Version bump only for package actionsflow